Created a python script to get serial data
We did this because of pythons capabilities to quickly connect to a serial device and receive data. In Java each new data event was taking way too long to process. I think the cause of this issue was the fact that we were using an event listener which has alot of code to run before the event and after it is processed.