Added CategoricalHashMap from DataAnalyzer.
CategoricalHashMap is a HashMap that uses seperate chaining on collisions. The HashCode is found from the LogObjects TAG. This class was created so that a TAG value could be given and a list of all the elements with that tag value would be returned. This reduces get() runtimes so that we do not have to iterate through all the elements just the ones we will use.